add call graph function
